fix(merge): uploads and resource in one panel
